Exclude all possible directions from default search
From now on, WordsearchBoard.exists() method uses only conventional set
of directions while searching for a word: horizontal, vertical, and diagonal.

There still a way to look for all possible directions by passing "all" as a
direction parameter value. Note that this would result in locating words
in mixed conditions which could not be a desired option.
See tests for details.